♻️ Refactor: mapper 분리
